Change file structure for api folder
